dolphin/requirements.txt
espnet==202402
modelscope
addict==2.4.0
pydub
funasr
torchvision
tqdm